refactor: pages directory
