+++ src/components/Organisms/AppBarComponent.js
@@ -3,8 +3,8 @@ import history from '../../utils/history'
 import LeftMenuBar from './LeftMenuBar'
 import Grid from '@material-ui/core/Grid'
 import React from 'react'
-import MainPage from './MainPage'
-import { HomePage } from './HomePage'
+import MainPage from '../Pages/MainPage'
+import { HomePage } from '../Pages/HomePage'
 import { EraseQuestionnaireButton } from '../Atoms/Buttons/EraseQuestionnaireButton'
 import { RenderQuestionnaireButton } from '../Atoms/Buttons/RenderQuestionnaireButton'
 import { SaveQuestionnaireButton } from '../Atoms/Buttons/SaveQuestionnaireButton'
